+++ src/scenegraph/Group.h
@@ -22,7 +22,7 @@ public:
 	virtual void AddChild(Node *child);
 	virtual bool RemoveChild(Node *node); //true on success
 	virtual bool RemoveChildAt(unsigned int position); //true on success
-	unsigned int GetNumChildren() const { return m_children.size(); }
+	unsigned int GetNumChildren() const { return static_cast<Uint32>(m_children.size()); }
 	Node* GetChildAt(unsigned int);
 	virtual void Accept(NodeVisitor &v) override;
 	virtual void Traverse(NodeVisitor &v) override;
