+++ src/scenegraph/StaticGeometry.h
@@ -38,7 +38,7 @@ public:
 	void AddMesh(RefCountedPtr<Graphics::VertexBuffer>,
 		RefCountedPtr<Graphics::IndexBuffer>,
 		RefCountedPtr<Graphics::Material>);
-	unsigned int GetNumMeshes() const { return m_meshes.size(); }
+	unsigned int GetNumMeshes() const { return static_cast<Uint32>(m_meshes.size()); }
 	Mesh &GetMeshAt(unsigned int i);
 
 	void SetRenderState(Graphics::RenderState *s) { m_renderState = s; }
